Add files via upload
I made new classes for the different vending machines and made hashes for the other food items. I started editing the menu so that the user can navigate between viewing current machines, adding new machines, and testing the machines. Right now we only have testing. This upload is mainly for saving the work and showing progress. I plan on getting the menu to send you to the different vending machine classes soon.